\name{ADMMImpute-package}
\alias{ADMMImpute-package}
\alias{ADMMImpute}
\docType{package}
\title{
  This package perform imputation using ADMM algorithm on scRNA-seq dataset
}
\description{
  This package perform imputation using ADMM algorithm on scRNA-seq dataset
}
\details{
  See function instructions for Impute. you can use ?Impute in R to get the instructions
}
\author{
Ying Ma, yingma@umich.edu.

Maintainer: Ying Ma <yingma@umich.edu.>
}
\references{
  See Li et al 2018 Nature Communications
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
