% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imputation_wlabel_model}
\alias{imputation_wlabel_model}
\title{A Rcpp function to perform imputation on scRNAseq data based on ADMM algorithm}
\usage{
imputation_wlabel_model(count, cell_labels, point, drop_thre, rho,
  max_iter, tol)
}
\arguments{
\item{count}{normalized count from raw data}

\item{cell_labels}{cell labels, it can be estimated or provided by user}

\item{point}{pseudo zero count}

\item{drop_thre}{dropout rate threshold}

\item{rho}{Step size in ADMM algprithm}

\item{max_iter}{max iteration in ADMM algorithm}

\item{tol}{tolerance}
}
\value{
A list containing imputed data
}
\description{
A Rcpp function to perform imputation on scRNAseq data based on ADMM algorithm
}
