% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{Impute}
\alias{Impute}
\title{A R function to apply imputation_model which was implemented in Rcpp functions to perform imputation on scRNA-seq data}
\usage{
Impute(raw_data, labeled = FALSE, labels = NULL, numCluster, drop_thre,
  rho = 10, max_iter = 5000, tol = 1e-04)
}
\arguments{
\item{raw_data}{The raw data in a matrix format you want to perform imputation on. And the rownames of the count matrix should representing the genes and column names should representing the cells.}

\item{labeled}{A logical value describing whether cell labels information are known.}

\item{labels}{A character vector describing the cell label information. Only needed when \code{labeled = TRUE}.}

\item{numCluster}{A number describing the expected or estimated number of subpopulations. This could be determined by prior knowledge or unsupervised clustering.}

\item{drop_thre}{A number between 0 and 1 describing the dropout rate threshold to determine the dropout genes and the non-dropout genes.}

\item{rho}{step size in ADMM algorithm, default value is 10.}

\item{max_iter}{max iterations in ADMM algorithm, default value is 5000}

\item{tol}{tolerance in ADMM algorithm, default value is 1e-4}
}
\value{
Impute returns imputed count matrix which matches the dimension of the original count matrix
}
\description{
A R function to apply imputation_model which was implemented in Rcpp functions to perform imputation on scRNA-seq data
}
